Fix not working delete

+++ Administration.Service.Data/Repositories/WorkersRepository.cs
@@ -92,30 +92,26 @@ namespace Administration.Service.Data.Repositories
 			};
 		}
 
-		public async Task<bool> RemoveWorkerServiceAsync(Guid serviceId, Guid workerId)
+		public async Task RemoveWorkerServiceAsync(Guid serviceId, Guid workerId)
 		{
-			var workerServiceToDelete = new WorkerService() { ServiceId = serviceId, UserId = workerId };
+			var workerServiceToDelete = await _dbContext.WorkerServices.FirstOrDefaultAsync(ws => ws.ServiceId == serviceId && ws.UserId == workerId);
 
-			// Attach the entity to the context and mark it as deleted - this would be less resource consuming than reading it from the DB
-			_dbContext.WorkerServices.Attach(workerServiceToDelete);
-			_dbContext.WorkerServices.Remove(workerServiceToDelete);
-
-			await _dbContext.SaveChangesAsync();
-
-			return true;
+			if (workerServiceToDelete != null)
+			{
+				_dbContext.WorkerServices.Remove(workerServiceToDelete);
+				await _dbContext.SaveChangesAsync();
+			}
 		}
 
-		public async Task<bool> RemoveSaloonWorkerAsync(Guid saloonId, Guid workerId)
+		public async Task RemoveSaloonWorkerAsync(Guid saloonId, Guid workerId)
 		{
-			var saloonWorkerToDelete = new SaloonWorker() { SaloonId = saloonId, UserId = workerId };
+			var saloonWorkerToDelete = await _dbContext.SaloonWorkers.FirstOrDefaultAsync(sw => sw.SaloonId == saloonId && sw.UserId == workerId);
 
-			// Attach the entity to the context and mark it as deleted - this would be less resource consuming than reading it from the DB
-			_dbContext.SaloonWorkers.Attach(saloonWorkerToDelete);
-			_dbContext.SaloonWorkers.Remove(saloonWorkerToDelete);
-
-			await _dbContext.SaveChangesAsync();
-
-			return true;
+			if (saloonWorkerToDelete != null)
+			{
+				_dbContext.SaloonWorkers.Remove(saloonWorkerToDelete);
+				await _dbContext.SaveChangesAsync();
+			}
 		}
 
 		public async Task DeleteWorkerAsync(Guid workerId)
